fix(project-gate): stop swallowing run-config parse failures; split manifest-reading into its own module
Tier-3 PR review on PR #729 found a deeper issue behind the previous two
rounds' run-config fallback fixes: read_run_config() swallows a
JSONDecodeError into {} by its own documented design, so
_declared_split_names could never distinguish a genuinely absent
shipwright_run_config.json from a present-but-syntactically-invalid one
— both prior shape checks operated only on whatever that helper handed
back, which could never carry a parse failure at all.
_declared_split_names now reads shipwright_run_config.json directly, the
same way it already reads shipwright_project_config.json, so a parse
failure surfaces its own manifest error.

Also split _is_safe_split_name, _declared_split_names, _read_spec_texts
and _unreadable_result out of _project_gate_wiring.py into a new sibling
module, _project_gate_manifest.py: this file had crossed the 300-line
bloat-baseline guideline on nearly every one of these Tier-3 fix rounds,
and further docstring trimming was starting to compress the review-history
context those docstrings exist to preserve.
